+++ recipes/sentinel_enable.rb
@@ -19,7 +19,8 @@
 #
 
 sentinel_instances = node['redisio']['sentinels']
-if sentinel_instances.empty?
+
+if sentinel_instances.nil?
   sentinel_instances = [{'sentinel_port' => '26379', 'name' => 'mycluster', 'master_ip' => '127.0.0.1', 'master_port' => '6379'}]
 end